backend/requirements.txt
deepface==0.0.93
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84